+++ packages/react-i18n/CHANGELOG.md
@@ -10,19 +10,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Add `unformatCurrency` utility return the normalized value of a formatted money value
+- Add `unformatCurrency` utility return the normalized value of a formatted money value [#486](https://github.com/Shopify/quilt/pull/486)
 
 ## [0.7.4] - 2019-01-21
 
 ### Added
 
-- `CurrencyCode` enum types [#473](https://github.com/Shopify/quilt/pull/473/files)
+- `CurrencyCode` enum types [#473](https://github.com/Shopify/quilt/pull/473)
 
 ## [0.7.3] - 2019-01-21
 
 ### Added
 
-- Format count when handling pluralized translation [#447](https://github.com/Shopify/quilt/pull/447/files)
+- Format count when handling pluralized translation [#447](https://github.com/Shopify/quilt/pull/447)
 
 ## [0.7.2]