+++ packages/react-i18n/CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ---
 
-## [Unreleased] -
+<!-- ## [Unreleased] - -->
+
+## [0.8.0] - 2019-01-30
 
 ### Added